Weight transactions (#48)
* Weight transaction models

* Fixed date migrations

* Weight services

* API endpoints for Weight Transactions

* Weight FE scaffolding

* Weigh transaction unit tests

* Weight state and unit test fixes

* Weight transaction form

The weight transaction form is a component that is focussed
on capturing the information relating to a weight transaction.
It will be hosted on both new and edit pages. It is a pure
component only concerned with rendering logic.

The pages, edit and new are located at routes:
- new: `/weight/new`
- edit: `/weight/:id/edit`
They are responsible for managing the data fetch requests
and async operations.