Allow archiving of animals
Animals can now be archived. The endpoint for fetching animals has been
updated to take in 3 sets of query parameters, paginationOptions,
orderingOptions and filter keys. Each have sub properties for paging,
ordering and filtering.

The AnimalSearchService has been moved away from the IFetchAsync
interface which has been deprecated an now makes use of SQL LINQ
expression extensions to allow for quick customization. This decouples
the services from having to implement a bunch of features that will
never be used.

The UI now provides a filter line on the animal table to toggle for
including and excluding archived items. It also allows for marking
animals as archived in bulk. Alternatively each animal can be archived
by viewing it individually.

Resolves: #62