Map DbContext to Domain models
To fix some seed data issues, all the EF Core model mappings now point
to the domain models. A lot of code has been discarded and a lot has
been reworked. This is essentially the start for ripping the feature
modules appart. No more unecessary interfaces and tightly coupled code.
Well, there is still more to do, but this is 90% better.